Remove Config.h include from many places
