+++ src/openrct2-ui/windows/TitleMenu.cpp
@@ -18,7 +18,6 @@
 #include <openrct2/ParkImporter.h>
 #include <openrct2/PlatformEnvironment.h>
 #include <openrct2/actions/LoadOrQuitAction.h>
-#include <openrct2/config/Config.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/sprites.h>
 #include <openrct2/ui/UiContext.h>
