+++ src/openrct2-ui/windows/TitleOptions.cpp
@@ -10,7 +10,6 @@
 #include <openrct2-ui/interface/Widget.h>
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Context.h>
-#include <openrct2/config/Config.h>
 #include <openrct2/localisation/Localisation.h>
 
 namespace OpenRCT2::Ui::Windows
