+++ src/openrct2/command_line/CommandLine.cpp
@@ -12,6 +12,7 @@
 #include "../OpenRCT2.h"
 #include "../core/Console.hpp"
 #include "../core/String.hpp"
+#include "../drawing/Font.h"
 #include "../platform/Platform.h"
 
 #include <cstring>
