+++ src/openrct2/interface/StdInOutConsole.cpp
@@ -11,6 +11,7 @@
 
 #include "../Context.h"
 #include "../OpenRCT2.h"
+#include "../config/ConfigTypes.h"
 #include "../platform/Platform.h"
 #include "../scripting/ScriptEngine.h"
