+++ src/openrct2/rct2/T6Importer.cpp
@@ -8,7 +8,6 @@
  *****************************************************************************/
 
 #include "../TrackImporter.h"
-#include "../config/Config.h"
 #include "../core/FileStream.h"
 #include "../core/MemoryStream.h"
 #include "../core/Path.hpp"
